+++ app/Http/Controllers/Auth/RegisterController.php
@@ -64,10 +64,23 @@ class RegisterController extends Controller
      */
     protected function create(array $data)
     {
-        return User::create([
+        /**
+         * Added for Sanctum API token creation on registration.
+         */
+
+        $_user = User::create([
             'name' => $data['name'],
             'email' => $data['email'],
             'password' => Hash::make($data['password']),
         ]);
+        $_user->createToken('HPDS-MAIN');
+        return $_user;
+
+        // Original create function.
+    //     return User::create([
+    //         'name' => $data['name'],
+    //         'email' => $data['email'],
+    //         'password' => Hash::make($data['password']),
+    //     ]);
     }
 }