+++ resources/js/components/Bible/PageComponent.vue
@@ -8,10 +8,21 @@
             <leaf-component class="col-1" type=-1 v-on:leaf-page="leafPage"></leaf-component>
             <div class="col-10">
                 <div class="verse" v-for="(text, index) in chapterText" :key="text.id">
-                    <p><span class="h5">{{ index + 1 }}: </span><span v-html="text.verse"></span></p>
-                    <div v-if="loggedIn == 'true'">
-                        <highlight-component></highlight-component>
+                    <!-- <div class="row" v-on:mouseover="toggleHighlightComponent(index)" v-on:mouseout="toggleHighlightComponent(index)"> -->
+                    <div class="row" v-on:click="showHighlightComponent(index)">
+                        <p :id=" 'verse-' + index" class="col-12"><span class="h5">{{ index + 1 }}: </span><span v-html="text.verse"></span></p>
+                        
+                        <div :id="'highlight-component-' + index" class="col-3 d-none">
+                            <div v-if="loggedIn == true">
+                                <highlight-component></highlight-component>
+                            </div>
+                            <div v-else>
+                                <a href="/login">Log in</a> to save verses.
+                            </div>
+                        </div>
+
                     </div>
+                    
                 </div>
             </div>
             
@@ -30,14 +41,56 @@
             'leaf-component': LeafComponent,
             'highlight-component': HighlightComponent
         },
+        created: function () {
+            
+        },
+        data: function () {
+            return {
+                highlightComponentIndex: null
+            };
+        },
         methods: {
+            hideHighlightComponent: function (_index) {
+                this.highlightComponentIndex = null;
+            },
             leafPage: function (_type){
                 this.$emit('leaf-page', _type);
+            },
+            showHighlightComponent: function (_index) {
+                if (this.highlightComponentIndex === null) {
+                    this.toggleHighlightComponent(_index);
+                    this.highlightComponentIndex = _index;
+                } else if (this.highlightComponentIndex !== _index) {
+                    this.toggleHighlightComponent(this.highlightComponentIndex);
+                    this.toggleHighlightComponent(_index);
+                    this.highlightComponentIndex = _index;
+                } else if (this.highlightComponentIndex === _index) {
+                    this.toggleHighlightComponent(_index);
+                    this.highlightComponentIndex = null;
+                }
+            },
+            toggleHighlightComponent: function (_index) {
+                let _verse = document.getElementById('verse-' + _index);
+                let _verseClasses = _verse.classList;
+                _verseClasses.toggle('col-9');
+                _verseClasses.toggle('col-12');
+
+                let _highlightComponent = document.getElementById("highlight-component-" + _index);
+                let _componentClasses = _highlightComponent.classList;
+                _componentClasses.toggle('d-none');
             }
         },
-        created: function () {
-            console.log(this.loggedIn);
+        mounted: function () {
+            console.log(this.apiToken);
+            axios.get('api/stest', {
+                headers: {
+                    'Authorization': `Bearer ${this.apiToken}`
+                }
+            })
+            .then(response => {
+                console.log(response);
+            });
         },
-        props: ['page', 'chapterText', 'loggedIn']
+        props: ['page', 'chapterText', 'loggedIn', 'apiToken']
     }
 </script>